Spotlight Initial Integration
I picked through the jQuery Spotlight code and added some integration
of this to switch navigation. This is very unpolished now and I hope to
fix it up in the next commit, but the results so far look promising. I
also cleaned up the project and deleted some unnecessary files.