Pre web navigation refactor
After experimenting with the spotlight feature, it has become apparent
that the code currently used to create the web navigation tree in
KB_Node.js and in inject.js will need to be rewritten. One, it feels
like the NavTree data structure code is too mixed with the
implementation. I’m not sure if scan and select should be controlled by
the node. Second, testing this commit on a google search resulted in
some large performance issues that I think must be related to applying
css to divs with large numbers of children or something.