Highlighting Progress
By studying and modifying the jQuery Spotlight plugin, I have created a
layered spotlight system that could be used to more naturally highlight
selected areas of the page. Hopefully this will be a simple integration
into the system, because project priorities are directed towards fixing
the keyboard and refactoring the navigation tree.